+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Don't send 0 as the response size for streaming responses, send nothing instead.
+
 ## [1.2.0] - 2022-01-31
 
 ### Added